A faster unique() function

Similar to unique, only optimized for working with date-time and some list columns.

funique(x)

Arguments

x
Input data. If x is not a data frame or date-time object, then it is simply passed to unique

Value

The unique values of x.

Examples

## create example data set
d <- data.frame(
  x = rnorm(1000),
  y = seq.POSIXt(as.POSIXct("2018-01-01"),
    as.POSIXct("2018-12-31"), length.out = 10)
)

## sample to create version with duplicates
dd <- d[c(1:1000, sample(1:1000, 500, replace = TRUE)), ]

## get only unique rows
funique(dd)

#>                 x                   y
#> 1    -1.400043517 2018-01-01 00:00:00
#> 2     0.255317055 2018-02-10 10:40:00
#> 3    -2.437263611 2018-03-22 22:20:00
#> 4    -0.005571287 2018-05-02 09:00:00
#> 5     0.621552721 2018-06-11 19:40:00
#> 6     1.148411606 2018-07-22 06:20:00
#> 7    -1.821817661 2018-08-31 17:00:00
#> 8    -0.247325302 2018-10-11 03:40:00
#> 9    -0.244199607 2018-11-20 13:20:00
#> 10   -0.282705449 2018-12-31 00:00:00
#> 11   -0.553699384 2018-01-01 00:00:00
#> 12    0.628982042 2018-02-10 10:40:00
#> 13    2.065024895 2018-03-22 22:20:00
#> 14   -1.630989402 2018-05-02 09:00:00
#> 15    0.512426950 2018-06-11 19:40:00
#> 16   -1.863011492 2018-07-22 06:20:00
#> 17   -0.522012515 2018-08-31 17:00:00
#> 18   -0.052601910 2018-10-11 03:40:00
#> 19    0.542996343 2018-11-20 13:20:00
#> 20   -0.914074827 2018-12-31 00:00:00
#> 21    0.468154420 2018-01-01 00:00:00
#> 22    0.362951256 2018-02-10 10:40:00
#> 23   -1.304543545 2018-03-22 22:20:00
#> 24    0.737776321 2018-05-02 09:00:00
#> 25    1.888504929 2018-06-11 19:40:00
#> 26   -0.097445104 2018-07-22 06:20:00
#> 27   -0.935847354 2018-08-31 17:00:00
#> 28   -0.015950311 2018-10-11 03:40:00
#> 29   -0.826788954 2018-11-20 13:20:00
#> 30   -1.512399651 2018-12-31 00:00:00
#> 31    0.935363190 2018-01-01 00:00:00
#> 32    0.176488611 2018-02-10 10:40:00
#> 33    0.243685465 2018-03-22 22:20:00
#> 34    1.623548883 2018-05-02 09:00:00
#> 35    0.112038083 2018-06-11 19:40:00
#> 36   -0.133997013 2018-07-22 06:20:00
#> 37   -1.910087468 2018-08-31 17:00:00
#> 38   -0.279237242 2018-10-11 03:40:00
#> 39   -0.313445978 2018-11-20 13:20:00
#> 40    1.067307879 2018-12-31 00:00:00
#> 41    0.070034850 2018-01-01 00:00:00
#> 42   -0.639123324 2018-02-10 10:40:00
#> 43   -0.049964899 2018-03-22 22:20:00
#> 44   -0.251483443 2018-05-02 09:00:00
#> 45    0.444797116 2018-06-11 19:40:00
#> 46    2.755417575 2018-07-22 06:20:00
#> 47    0.046531380 2018-08-31 17:00:00
#> 48    0.577709069 2018-10-11 03:40:00
#> 49    0.118194874 2018-11-20 13:20:00
#> 50   -1.911720491 2018-12-31 00:00:00
#> 51    0.862086482 2018-01-01 00:00:00
#> 52   -0.243236740 2018-02-10 10:40:00
#> 53   -0.206087195 2018-03-22 22:20:00
#> 54    0.019177592 2018-05-02 09:00:00
#> 55    0.029560754 2018-06-11 19:40:00
#> 56    0.549827542 2018-07-22 06:20:00
#> 57   -2.274114857 2018-08-31 17:00:00
#> 58    2.682557184 2018-10-11 03:40:00
#> 59   -0.361221255 2018-11-20 13:20:00
#> 60    0.213355750 2018-12-31 00:00:00
#> 61    1.074345882 2018-01-01 00:00:00
#> 62   -0.665088249 2018-02-10 10:40:00
#> 63    1.113952419 2018-03-22 22:20:00
#> 64   -0.245896412 2018-05-02 09:00:00
#> 65   -1.177563309 2018-06-11 19:40:00
#> 66   -0.975850616 2018-07-22 06:20:00
#> 67    1.065057320 2018-08-31 17:00:00
#> 68    0.131670635 2018-10-11 03:40:00
#> 69    0.488628809 2018-11-20 13:20:00
#> 70   -1.699450568 2018-12-31 00:00:00
#> 71   -1.470736306 2018-01-01 00:00:00
#> 72    0.284150344 2018-02-10 10:40:00
#> 73    1.337320413 2018-03-22 22:20:00
#> 74    0.236696283 2018-05-02 09:00:00
#> 75    1.318293384 2018-06-11 19:40:00
#> 76    0.523909788 2018-07-22 06:20:00
#> 77    0.606748047 2018-08-31 17:00:00
#> 78   -0.109935672 2018-10-11 03:40:00
#> 79    0.172181715 2018-11-20 13:20:00
#> 80   -0.090327287 2018-12-31 00:00:00
#> 81    1.924343341 2018-01-01 00:00:00
#> 82    1.298392759 2018-02-10 10:40:00
#> 83    0.748791268 2018-03-22 22:20:00
#> 84    0.556224329 2018-05-02 09:00:00
#> 85   -0.548257264 2018-06-11 19:40:00
#> 86    1.110534893 2018-07-22 06:20:00
#> 87   -2.612334333 2018-08-31 17:00:00
#> 88   -0.155693776 2018-10-11 03:40:00
#> 89    0.433889790 2018-11-20 13:20:00
#> 90   -0.381951112 2018-12-31 00:00:00
#> 91    0.424187575 2018-01-01 00:00:00
#> 92    1.063101996 2018-02-10 10:40:00
#> 93    1.048712620 2018-03-22 22:20:00
#> 94   -0.038102895 2018-05-02 09:00:00
#> 95    0.486148920 2018-06-11 19:40:00
#> 96    1.672882611 2018-07-22 06:20:00
#> 97   -0.354361164 2018-08-31 17:00:00
#> 98    0.946347886 2018-10-11 03:40:00
#> 99    1.316826356 2018-11-20 13:20:00
#> 100  -0.296640025 2018-12-31 00:00:00
#> 101  -0.387213575 2018-01-01 00:00:00
#> 102  -0.785432656 2018-02-10 10:40:00
#> 103  -1.056736867 2018-03-22 22:20:00
#> 104  -0.795541430 2018-05-02 09:00:00
#> 105  -1.756275428 2018-06-11 19:40:00
#> 106  -0.690537897 2018-07-22 06:20:00
#> 107  -0.558541994 2018-08-31 17:00:00
#> 108  -0.536663326 2018-10-11 03:40:00
#> 109   0.227127133 2018-11-20 13:20:00
#> 110   0.978454920 2018-12-31 00:00:00
#> 111  -0.208882651 2018-01-01 00:00:00
#> 112  -1.399410460 2018-02-10 10:40:00
#> 113   0.258537288 2018-03-22 22:20:00
#> 114  -0.441799453 2018-05-02 09:00:00
#> 115   0.568599861 2018-06-11 19:40:00
#> 116   2.126850459 2018-07-22 06:20:00
#> 117   0.424858441 2018-08-31 17:00:00
#> 118  -1.684281532 2018-10-11 03:40:00
#> 119   0.249401784 2018-11-20 13:20:00
#> 120   1.072838252 2018-12-31 00:00:00
#> 121   2.039369263 2018-01-01 00:00:00
#> 122   0.449453778 2018-02-10 10:40:00
#> 123   1.391814046 2018-03-22 22:20:00
#> 124   0.426566547 2018-05-02 09:00:00
#> 125   0.107583992 2018-06-11 19:40:00
#> 126   0.022294733 2018-07-22 06:20:00
#> 127   0.603611011 2018-08-31 17:00:00
#> 128  -0.262650573 2018-10-11 03:40:00
#> 129  -0.528264082 2018-11-20 13:20:00
#> 130   0.192149422 2018-12-31 00:00:00
#> 131  -1.146199669 2018-01-01 00:00:00
#> 132   0.846184665 2018-02-10 10:40:00
#> 133   0.081719629 2018-03-22 22:20:00
#> 134  -1.305117010 2018-05-02 09:00:00
#> 135  -0.944912060 2018-06-11 19:40:00
#> 136   0.454341594 2018-07-22 06:20:00
#> 137  -0.855202501 2018-08-31 17:00:00
#> 138  -0.286895219 2018-10-11 03:40:00
#> 139   0.894961626 2018-11-20 13:20:00
#> 140   0.067304440 2018-12-31 00:00:00
#> 141  -0.162676337 2018-01-01 00:00:00
#> 142  -0.827310169 2018-02-10 10:40:00
#> 143   1.876505621 2018-03-22 22:20:00
#> 144   0.766440199 2018-05-02 09:00:00
#> 145   0.979956696 2018-06-11 19:40:00
#> 146   1.321780992 2018-07-22 06:20:00
#> 147  -1.119710829 2018-08-31 17:00:00
#> 148   0.514599819 2018-10-11 03:40:00
#> 149  -1.509099836 2018-11-20 13:20:00
#> 150   1.532741480 2018-12-31 00:00:00
#> 151   0.429147371 2018-01-01 00:00:00
#> 152   0.122103414 2018-02-10 10:40:00
#> 153  -1.138012401 2018-03-22 22:20:00
#> 154  -0.558015129 2018-05-02 09:00:00
#> 155   1.052538537 2018-06-11 19:40:00
#> 156   0.677683644 2018-07-22 06:20:00
#> 157   0.038499547 2018-08-31 17:00:00
#> 158  -0.356381187 2018-10-11 03:40:00
#> 159   0.782844102 2018-11-20 13:20:00
#> 160   0.804411616 2018-12-31 00:00:00
#> 161  -1.900060823 2018-01-01 00:00:00
#> 162   0.935784286 2018-02-10 10:40:00
#> 163  -0.309051503 2018-03-22 22:20:00
#> 164   0.263066677 2018-05-02 09:00:00
#> 165  -1.790591856 2018-06-11 19:40:00
#> 166  -0.788258845 2018-07-22 06:20:00
#> 167  -1.133021669 2018-08-31 17:00:00
#> 168   0.363652568 2018-10-11 03:40:00
#> 169  -0.285887914 2018-11-20 13:20:00
#> 170   0.517669134 2018-12-31 00:00:00
#> 171  -0.102908670 2018-01-01 00:00:00
#> 172  -0.974069593 2018-02-10 10:40:00
#> 173   1.270672301 2018-03-22 22:20:00
#> 174   0.960864787 2018-05-02 09:00:00
#> 175   0.768721370 2018-06-11 19:40:00
#> 176   1.035930771 2018-07-22 06:20:00
#> 177  -0.473887074 2018-08-31 17:00:00
#> 178  -1.275334875 2018-10-11 03:40:00
#> 179  -0.305620674 2018-11-20 13:20:00
#> 180   2.211769487 2018-12-31 00:00:00
#> 181  -1.041668381 2018-01-01 00:00:00
#> 182  -1.146523850 2018-02-10 10:40:00
#> 183  -1.675327303 2018-03-22 22:20:00
#> 184   1.525938655 2018-05-02 09:00:00
#> 185   0.554185515 2018-06-11 19:40:00
#> 186   1.993110265 2018-07-22 06:20:00
#> 187  -0.154120740 2018-08-31 17:00:00
#> 188   2.564408338 2018-10-11 03:40:00
#> 189   1.061999145 2018-11-20 13:20:00
#> 190   1.142694878 2018-12-31 00:00:00
#> 191   1.123838843 2018-01-01 00:00:00
#> 192  -0.397001493 2018-02-10 10:40:00
#> 193  -0.823261151 2018-03-22 22:20:00
#> 194  -0.578884625 2018-05-02 09:00:00
#> 195   1.763789378 2018-06-11 19:40:00
#> 196   0.132992146 2018-07-22 06:20:00
#> 197   0.376499328 2018-08-31 17:00:00
#> 198   1.138707653 2018-10-11 03:40:00
#> 199   1.241263075 2018-11-20 13:20:00
#> 200   0.612090945 2018-12-31 00:00:00
#> 201  -0.429380087 2018-01-01 00:00:00
#> 202   1.360461327 2018-02-10 10:40:00
#> 203  -0.070857431 2018-03-22 22:20:00
#> 204  -0.272153684 2018-05-02 09:00:00
#> 205  -2.446680029 2018-06-11 19:40:00
#> 206   0.065486641 2018-07-22 06:20:00
#> 207  -1.098508902 2018-08-31 17:00:00
#> 208  -0.633178176 2018-10-11 03:40:00
#> 209  -2.063654451 2018-11-20 13:20:00
#> 210   2.648932029 2018-12-31 00:00:00
#> 211  -1.153398386 2018-01-01 00:00:00
#> 212  -0.340637876 2018-02-10 10:40:00
#> 213   0.786362576 2018-03-22 22:20:00
#> 214  -1.270513110 2018-05-02 09:00:00
#> 215   0.542141549 2018-06-11 19:40:00
#> 216   0.075105900 2018-07-22 06:20:00
#> 217   0.558514422 2018-08-31 17:00:00
#> 218   0.415406399 2018-10-11 03:40:00
#> 219  -1.452299769 2018-11-20 13:20:00
#> 220   0.941206122 2018-12-31 00:00:00
#> 221  -0.338935872 2018-01-01 00:00:00
#> 222  -0.075574247 2018-02-10 10:40:00
#> 223   0.040204392 2018-03-22 22:20:00
#> 224   0.124301066 2018-05-02 09:00:00
#> 225  -0.998432551 2018-06-11 19:40:00
#> 226   1.233390065 2018-07-22 06:20:00
#> 227   0.340424488 2018-08-31 17:00:00
#> 228  -0.472702482 2018-10-11 03:40:00
#> 229   0.708753061 2018-11-20 13:20:00
#> 230  -1.528958715 2018-12-31 00:00:00
#> 231   0.237425345 2018-01-01 00:00:00
#> 232  -1.312814246 2018-02-10 10:40:00
#> 233   0.747028587 2018-03-22 22:20:00
#> 234  -1.562518435 2018-05-02 09:00:00
#> 235   0.071053360 2018-06-11 19:40:00
#> 236  -0.639534770 2018-07-22 06:20:00
#> 237  -0.845195739 2018-08-31 17:00:00
#> 238   0.675244698 2018-10-11 03:40:00
#> 239   1.153375794 2018-11-20 13:20:00
#> 240  -1.686504742 2018-12-31 00:00:00
#> 241  -0.902814949 2018-01-01 00:00:00
#> 242   1.317633698 2018-02-10 10:40:00
#> 243   1.100189745 2018-03-22 22:20:00
#> 244   1.203767839 2018-05-02 09:00:00
#> 245  -1.431270777 2018-06-11 19:40:00
#> 246   1.382910861 2018-07-22 06:20:00
#> 247   0.003125940 2018-08-31 17:00:00
#> 248  -0.077886824 2018-10-11 03:40:00
#> 249   0.441428226 2018-11-20 13:20:00
#> 250   0.128922896 2018-12-31 00:00:00
#> 251  -0.830214260 2018-01-01 00:00:00
#> 252  -0.503592910 2018-02-10 10:40:00
#> 253  -1.193641182 2018-03-22 22:20:00
#> 254  -0.751723323 2018-05-02 09:00:00
#> 255   1.455841403 2018-06-11 19:40:00
#> 256  -0.828603533 2018-07-22 06:20:00
#> 257   0.289774460 2018-08-31 17:00:00
#> 258  -0.480053484 2018-10-11 03:40:00
#> 259  -0.604829354 2018-11-20 13:20:00
#> 260   1.460110180 2018-12-31 00:00:00
#> 261   0.149679354 2018-01-01 00:00:00
#> 262  -1.433321100 2018-02-10 10:40:00
#> 263  -0.010303319 2018-03-22 22:20:00
#> 264  -0.212236035 2018-05-02 09:00:00
#> 265  -0.906340179 2018-06-11 19:40:00
#> 266  -2.102152479 2018-07-22 06:20:00
#> 267   1.893360464 2018-08-31 17:00:00
#> 268  -0.968125837 2018-10-11 03:40:00
#> 269  -0.102603036 2018-11-20 13:20:00
#> 270   0.239959572 2018-12-31 00:00:00
#> 271   0.060898893 2018-01-01 00:00:00
#> 272  -2.177576028 2018-02-10 10:40:00
#> 273  -0.117860143 2018-03-22 22:20:00
#> 274   0.112294787 2018-05-02 09:00:00
#> 275   0.007886198 2018-06-11 19:40:00
#> 276   1.877743872 2018-07-22 06:20:00
#> 277   2.158756554 2018-08-31 17:00:00
#> 278   0.709714522 2018-10-11 03:40:00
#> 279   0.766983379 2018-11-20 13:20:00
#> 280  -0.308211421 2018-12-31 00:00:00
#> 281   1.012001849 2018-01-01 00:00:00
#> 282  -0.919051597 2018-02-10 10:40:00
#> 283   0.563380077 2018-03-22 22:20:00
#> 284   0.322482749 2018-05-02 09:00:00
#> 285   0.366674363 2018-06-11 19:40:00
#> 286   1.129835153 2018-07-22 06:20:00
#> 287  -0.941498076 2018-08-31 17:00:00
#> 288   0.217837643 2018-10-11 03:40:00
#> 289   1.415412293 2018-11-20 13:20:00
#> 290  -0.383733048 2018-12-31 00:00:00
#> 291  -0.174086374 2018-01-01 00:00:00
#> 292  -0.221744517 2018-02-10 10:40:00
#> 293  -1.009528722 2018-03-22 22:20:00
#> 294   0.480725266 2018-05-02 09:00:00
#> 295   1.604407328 2018-06-11 19:40:00
#> 296  -1.515024529 2018-07-22 06:20:00
#> 297  -1.416023914 2018-08-31 17:00:00
#> 298   0.876777327 2018-10-11 03:40:00
#> 299   0.624132413 2018-11-20 13:20:00
#> 300   2.112277288 2018-12-31 00:00:00
#> 301  -0.356124416 2018-01-01 00:00:00
#> 302  -1.064464209 2018-02-10 10:40:00
#> 303   1.077116538 2018-03-22 22:20:00
#> 304   1.181575567 2018-05-02 09:00:00
#> 305   0.198392095 2018-06-11 19:40:00
#> 306  -0.400405249 2018-07-22 06:20:00
#> 307   0.616154281 2018-08-31 17:00:00
#> 308   1.974156748 2018-10-11 03:40:00
#> 309   1.884662324 2018-11-20 13:20:00
#> 310  -1.588620547 2018-12-31 00:00:00
#> 311  -0.539923164 2018-01-01 00:00:00
#> 312  -1.169461464 2018-02-10 10:40:00
#> 313   0.559105989 2018-03-22 22:20:00
#> 314  -1.819347247 2018-05-02 09:00:00
#> 315   0.393343972 2018-06-11 19:40:00
#> 316   0.042134106 2018-07-22 06:20:00
#> 317   1.179664177 2018-08-31 17:00:00
#> 318  -0.256921176 2018-10-11 03:40:00
#> 319  -1.056336098 2018-11-20 13:20:00
#> 320   0.198777205 2018-12-31 00:00:00
#> 321   0.650533552 2018-01-01 00:00:00
#> 322   0.343913337 2018-02-10 10:40:00
#> 323   1.477532312 2018-03-22 22:20:00
#> 324   0.072025698 2018-05-02 09:00:00
#> 325   2.126444534 2018-06-11 19:40:00
#> 326  -1.476196906 2018-07-22 06:20:00
#> 327   0.407888500 2018-08-31 17:00:00
#> 328   1.393977798 2018-10-11 03:40:00
#> 329   0.360278296 2018-11-20 13:20:00
#> 330   0.654550251 2018-12-31 00:00:00
#> 331   1.052155422 2018-01-01 00:00:00
#> 332  -1.979555125 2018-02-10 10:40:00
#> 333   1.208385605 2018-03-22 22:20:00
#> 334  -0.169280084 2018-05-02 09:00:00
#> 335   0.295029753 2018-06-11 19:40:00
#> 336   1.266340587 2018-07-22 06:20:00
#> 337  -1.135343257 2018-08-31 17:00:00
#> 338  -1.131053798 2018-10-11 03:40:00
#> 339   0.109993384 2018-11-20 13:20:00
#> 340   0.852905410 2018-12-31 00:00:00
#> 341  -0.234337862 2018-01-01 00:00:00
#> 342   2.086688567 2018-02-10 10:40:00
#> 343  -0.110919371 2018-03-22 22:20:00
#> 344  -1.392847056 2018-05-02 09:00:00
#> 345  -1.142290768 2018-06-11 19:40:00
#> 346   1.704608737 2018-07-22 06:20:00
#> 347  -0.080073634 2018-08-31 17:00:00
#> 348  -0.437281240 2018-10-11 03:40:00
#> 349  -0.119215094 2018-11-20 13:20:00
#> 350   0.786462865 2018-12-31 00:00:00
#> 351  -0.578945246 2018-01-01 00:00:00
#> 352  -0.145426885 2018-02-10 10:40:00
#> 353   0.526457991 2018-03-22 22:20:00
#> 354   1.733578110 2018-05-02 09:00:00
#> 355   1.448657220 2018-06-11 19:40:00
#> 356   1.518193149 2018-07-22 06:20:00
#> 357  -0.384007254 2018-08-31 17:00:00
#> 358   1.827125177 2018-10-11 03:40:00
#> 359  -0.551491750 2018-11-20 13:20:00
#> 360  -0.865753541 2018-12-31 00:00:00
#> 361  -0.343831481 2018-01-01 00:00:00
#> 362   1.062876458 2018-02-10 10:40:00
#> 363   0.813058204 2018-03-22 22:20:00
#> 364   1.803483361 2018-05-02 09:00:00
#> 365  -0.105068694 2018-06-11 19:40:00
#> 366   0.982453362 2018-07-22 06:20:00
#> 367  -1.713302622 2018-08-31 17:00:00
#> 368  -0.832019527 2018-10-11 03:40:00
#> 369   1.100491882 2018-11-20 13:20:00
#> 370  -0.173820106 2018-12-31 00:00:00
#> 371   0.178812018 2018-01-01 00:00:00
#> 372  -0.698429449 2018-02-10 10:40:00
#> 373  -0.960449159 2018-03-22 22:20:00
#> 374  -0.975423042 2018-05-02 09:00:00
#> 375  -0.338576503 2018-06-11 19:40:00
#> 376   1.152347074 2018-07-22 06:20:00
#> 377   0.405101183 2018-08-31 17:00:00
#> 378  -0.470922497 2018-10-11 03:40:00
#> 379  -0.133251019 2018-11-20 13:20:00
#> 380   1.226682356 2018-12-31 00:00:00
#> 381   0.332943995 2018-01-01 00:00:00
#> 382  -0.347088466 2018-02-10 10:40:00
#> 383  -0.098550690 2018-03-22 22:20:00
#> 384   0.034766060 2018-05-02 09:00:00
#> 385   0.386127022 2018-06-11 19:40:00
#> 386   0.020831228 2018-07-22 06:20:00
#> 387   0.007586777 2018-08-31 17:00:00
#> 388   0.930844030 2018-10-11 03:40:00
#> 389  -0.684749941 2018-11-20 13:20:00
#> 390   0.337401513 2018-12-31 00:00:00
#> 391  -0.412137704 2018-01-01 00:00:00
#> 392   0.934261130 2018-02-10 10:40:00
#> 393   1.840316741 2018-03-22 22:20:00
#> 394  -0.704819663 2018-05-02 09:00:00
#> 395   0.008510312 2018-06-11 19:40:00
#> 396   2.034189886 2018-07-22 06:20:00
#> 397  -1.341686068 2018-08-31 17:00:00
#> 398   1.158979182 2018-10-11 03:40:00
#> 399  -0.203208958 2018-11-20 13:20:00
#> 400  -0.378028555 2018-12-31 00:00:00
#> 401   1.736111043 2018-01-01 00:00:00
#> 402  -0.845247816 2018-02-10 10:40:00
#> 403  -0.961571493 2018-03-22 22:20:00
#> 404   1.017491053 2018-05-02 09:00:00
#> 405  -1.496053742 2018-06-11 19:40:00
#> 406  -1.184818730 2018-07-22 06:20:00
#> 407   0.630234373 2018-08-31 17:00:00
#> 408   2.101252514 2018-10-11 03:40:00
#> 409  -0.613736810 2018-11-20 13:20:00
#> 410  -1.634638272 2018-12-31 00:00:00
#> 411  -0.010441117 2018-01-01 00:00:00
#> 412  -0.656506139 2018-02-10 10:40:00
#> 413  -0.669533441 2018-03-22 22:20:00
#> 414  -0.478589028 2018-05-02 09:00:00
#> 415   1.319456316 2018-06-11 19:40:00
#> 416   0.636562761 2018-07-22 06:20:00
#> 417   0.514327782 2018-08-31 17:00:00
#> 418  -1.751375113 2018-10-11 03:40:00
#> 419   0.893597518 2018-11-20 13:20:00
#> 420   0.223038372 2018-12-31 00:00:00
#> 421   0.580816593 2018-01-01 00:00:00
#> 422  -0.177821421 2018-02-10 10:40:00
#> 423   0.740966708 2018-03-22 22:20:00
#> 424  -0.997443079 2018-05-02 09:00:00
#> 425  -2.938977561 2018-06-11 19:40:00
#> 426   0.719015661 2018-07-22 06:20:00
#> 427  -0.698005041 2018-08-31 17:00:00
#> 428  -1.894125843 2018-10-11 03:40:00
#> 429   0.076299249 2018-11-20 13:20:00
#> 430   0.875308501 2018-12-31 00:00:00
#> 431   0.453827393 2018-01-01 00:00:00
#> 432  -0.850716906 2018-02-10 10:40:00
#> 433   0.566201613 2018-03-22 22:20:00
#> 434   1.152211954 2018-05-02 09:00:00
#> 435  -0.756197377 2018-06-11 19:40:00
#> 436  -0.489258334 2018-07-22 06:20:00
#> 437  -1.166052337 2018-08-31 17:00:00
#> 438  -0.479668950 2018-10-11 03:40:00
#> 439   0.115348218 2018-11-20 13:20:00
#> 440  -1.768048407 2018-12-31 00:00:00
#> 441  -1.407638919 2018-01-01 00:00:00
#> 442   0.709178461 2018-02-10 10:40:00
#> 443  -1.240842940 2018-03-22 22:20:00
#> 444  -0.368327348 2018-05-02 09:00:00
#> 445   0.462080093 2018-06-11 19:40:00
#> 446  -0.322833101 2018-07-22 06:20:00
#> 447  -1.287214810 2018-08-31 17:00:00
#> 448  -1.030040247 2018-10-11 03:40:00
#> 449   1.514089316 2018-11-20 13:20:00
#> 450   0.346903586 2018-12-31 00:00:00
#> 451   1.779441542 2018-01-01 00:00:00
#> 452   0.386630924 2018-02-10 10:40:00
#> 453  -0.918695239 2018-03-22 22:20:00
#> 454  -1.584336488 2018-05-02 09:00:00
#> 455  -0.084058892 2018-06-11 19:40:00
#> 456  -2.085070889 2018-07-22 06:20:00
#> 457   0.003567992 2018-08-31 17:00:00
#> 458  -0.355770822 2018-10-11 03:40:00
#> 459   1.146359751 2018-11-20 13:20:00
#> 460  -0.221188446 2018-12-31 00:00:00
#> 461   1.018179021 2018-01-01 00:00:00
#> 462  -0.263719295 2018-02-10 10:40:00
#> 463   1.658542305 2018-03-22 22:20:00
#> 464  -0.774086771 2018-05-02 09:00:00
#> 465  -0.923937880 2018-06-11 19:40:00
#> 466  -0.275533378 2018-07-22 06:20:00
#> 467  -0.593399688 2018-08-31 17:00:00
#> 468  -0.122285891 2018-10-11 03:40:00
#> 469   1.179784246 2018-11-20 13:20:00
#> 470   0.641037374 2018-12-31 00:00:00
#> 471  -0.629588508 2018-01-01 00:00:00
#> 472  -0.807734971 2018-02-10 10:40:00
#> 473  -0.860489929 2018-03-22 22:20:00
#> 474  -2.169238693 2018-05-02 09:00:00
#> 475  -1.375836518 2018-06-11 19:40:00
#> 476  -0.493132472 2018-07-22 06:20:00
#> 477  -0.581652027 2018-08-31 17:00:00
#> 478  -0.167229304 2018-10-11 03:40:00
#> 479   0.485993129 2018-11-20 13:20:00
#> 480  -1.333395796 2018-12-31 00:00:00
#> 481  -0.261965625 2018-01-01 00:00:00
#> 482   0.652386303 2018-02-10 10:40:00
#> 483   0.748854971 2018-03-22 22:20:00
#> 484   0.896560285 2018-05-02 09:00:00
#> 485   1.489300424 2018-06-11 19:40:00
#> 486  -0.659403481 2018-07-22 06:20:00
#> 487   0.537283179 2018-08-31 17:00:00
#> 488   0.746803067 2018-10-11 03:40:00
#> 489   1.896317084 2018-11-20 13:20:00
#> 490  -2.060070725 2018-12-31 00:00:00
#> 491   0.064543870 2018-01-01 00:00:00
#> 492  -0.265147403 2018-02-10 10:40:00
#> 493  -0.447344531 2018-03-22 22:20:00
#> 494  -1.410700927 2018-05-02 09:00:00
#> 495  -0.506418882 2018-06-11 19:40:00
#> 496  -0.269761838 2018-07-22 06:20:00
#> 497  -1.085154918 2018-08-31 17:00:00
#> 498   0.362159127 2018-10-11 03:40:00
#> 499  -0.335672143 2018-11-20 13:20:00
#> 500   1.363804498 2018-12-31 00:00:00
#> 501  -0.711524136 2018-01-01 00:00:00
#> 502   0.662178797 2018-02-10 10:40:00
#> 503   0.291130223 2018-03-22 22:20:00
#> 504   0.197958000 2018-05-02 09:00:00
#> 505  -1.203566106 2018-06-11 19:40:00
#> 506  -0.039817044 2018-07-22 06:20:00
#> 507   0.686982465 2018-08-31 17:00:00
#> 508   0.705267007 2018-10-11 03:40:00
#> 509   0.991441680 2018-11-20 13:20:00
#> 510   1.144248971 2018-12-31 00:00:00
#> 511  -1.238910243 2018-01-01 00:00:00
#> 512   2.654898333 2018-02-10 10:40:00
#> 513  -0.156917189 2018-03-22 22:20:00
#> 514  -0.423490117 2018-05-02 09:00:00
#> 515  -0.198387058 2018-06-11 19:40:00
#> 516  -0.894802407 2018-07-22 06:20:00
#> 517   0.904269119 2018-08-31 17:00:00
#> 518   0.079649210 2018-10-11 03:40:00
#> 519  -1.258827223 2018-11-20 13:20:00
#> 520   1.025685106 2018-12-31 00:00:00
#> 521  -0.730778603 2018-01-01 00:00:00
#> 522  -0.190145507 2018-02-10 10:40:00
#> 523   0.528864693 2018-03-22 22:20:00
#> 524   0.550210535 2018-05-02 09:00:00
#> 525   0.549684337 2018-06-11 19:40:00
#> 526  -0.659542372 2018-07-22 06:20:00
#> 527   0.057421706 2018-08-31 17:00:00
#> 528  -2.808010508 2018-10-11 03:40:00
#> 529  -0.912259753 2018-11-20 13:20:00
#> 530  -0.782379163 2018-12-31 00:00:00
#> 531  -0.664104924 2018-01-01 00:00:00
#> 532   0.626309770 2018-02-10 10:40:00
#> 533  -0.507248206 2018-03-22 22:20:00
#> 534   0.270361335 2018-05-02 09:00:00
#> 535   0.467476865 2018-06-11 19:40:00
#> 536   0.723994958 2018-07-22 06:20:00
#> 537   0.613836939 2018-08-31 17:00:00
#> 538  -0.617869202 2018-10-11 03:40:00
#> 539   0.220724902 2018-11-20 13:20:00
#> 540   1.127926598 2018-12-31 00:00:00
#> 541   1.813454336 2018-01-01 00:00:00
#> 542  -0.083825685 2018-02-10 10:40:00
#> 543   1.367706666 2018-03-22 22:20:00
#> 544  -0.627434620 2018-05-02 09:00:00
#> 545  -0.216629150 2018-06-11 19:40:00
#> 546  -0.683713824 2018-07-22 06:20:00
#> 547  -0.444702734 2018-08-31 17:00:00
#> 548   0.606489806 2018-10-11 03:40:00
#> 549   0.624183075 2018-11-20 13:20:00
#> 550  -0.695431074 2018-12-31 00:00:00
#> 551  -0.783639078 2018-01-01 00:00:00
#> 552  -0.953123859 2018-02-10 10:40:00
#> 553   1.792756071 2018-03-22 22:20:00
#> 554   0.348976696 2018-05-02 09:00:00
#> 555   0.259103768 2018-06-11 19:40:00
#> 556  -0.805951897 2018-07-22 06:20:00
#> 557   0.105664701 2018-08-31 17:00:00
#> 558  -0.333599682 2018-10-11 03:40:00
#> 559   1.641847970 2018-11-20 13:20:00
#> 560  -0.643905859 2018-12-31 00:00:00
#> 561   0.587020562 2018-01-01 00:00:00
#> 562  -0.150403088 2018-02-10 10:40:00
#> 563  -1.710821848 2018-03-22 22:20:00
#> 564   1.431032558 2018-05-02 09:00:00
#> 565  -2.645212268 2018-06-11 19:40:00
#> 566  -1.032457405 2018-07-22 06:20:00
#> 567  -0.707466431 2018-08-31 17:00:00
#> 568  -0.700560014 2018-10-11 03:40:00
#> 569   0.537885439 2018-11-20 13:20:00
#> 570  -0.316332175 2018-12-31 00:00:00
#> 571  -0.839622754 2018-01-01 00:00:00
#> 572  -1.354928062 2018-02-10 10:40:00
#> 573  -0.817568272 2018-03-22 22:20:00
#> 574  -0.634400003 2018-05-02 09:00:00
#> 575   0.815949433 2018-06-11 19:40:00
#> 576   0.302795706 2018-07-22 06:20:00
#> 577   1.807086625 2018-08-31 17:00:00
#> 578  -0.894026756 2018-10-11 03:40:00
#> 579  -0.046428211 2018-11-20 13:20:00
#> 580  -0.471179138 2018-12-31 00:00:00
#> 581  -0.526692630 2018-01-01 00:00:00
#> 582  -0.095134908 2018-02-10 10:40:00
#> 583  -2.495364809 2018-03-22 22:20:00
#> 584   0.166889217 2018-05-02 09:00:00
#> 585   0.350492384 2018-06-11 19:40:00
#> 586   1.433701009 2018-07-22 06:20:00
#> 587   0.765906803 2018-08-31 17:00:00
#> 588   1.167520670 2018-10-11 03:40:00
#> 589  -0.136943429 2018-11-20 13:20:00
#> 590  -0.514902044 2018-12-31 00:00:00
#> 591   1.519744468 2018-01-01 00:00:00
#> 592  -0.328491678 2018-02-10 10:40:00
#> 593  -0.053671506 2018-03-22 22:20:00
#> 594  -0.563524635 2018-05-02 09:00:00
#> 595  -0.743908963 2018-06-11 19:40:00
#> 596  -0.109041651 2018-07-22 06:20:00
#> 597  -0.560829227 2018-08-31 17:00:00
#> 598   0.188001549 2018-10-11 03:40:00
#> 599   0.748850942 2018-11-20 13:20:00
#> 600  -1.916538316 2018-12-31 00:00:00
#> 601   0.236095847 2018-01-01 00:00:00
#> 602   0.628953415 2018-02-10 10:40:00
#> 603   0.417925676 2018-03-22 22:20:00
#> 604   1.976758477 2018-05-02 09:00:00
#> 605  -0.506286298 2018-06-11 19:40:00
#> 606  -1.109968853 2018-07-22 06:20:00
#> 607  -0.948705723 2018-08-31 17:00:00
#> 608   0.476843757 2018-10-11 03:40:00
#> 609  -0.795201560 2018-11-20 13:20:00
#> 610   0.234326923 2018-12-31 00:00:00
#> 611  -1.222451097 2018-01-01 00:00:00
#> 612  -2.453647354 2018-02-10 10:40:00
#> 613  -1.489260814 2018-03-22 22:20:00
#> 614  -0.432147734 2018-05-02 09:00:00
#> 615  -0.942554006 2018-06-11 19:40:00
#> 616  -0.121450799 2018-07-22 06:20:00
#> 617   1.336446798 2018-08-31 17:00:00
#> 618  -0.860356182 2018-10-11 03:40:00
#> 619   0.666537820 2018-11-20 13:20:00
#> 620  -1.421534746 2018-12-31 00:00:00
#> 621   1.170056168 2018-01-01 00:00:00
#> 622  -1.404714543 2018-02-10 10:40:00
#> 623   1.101708096 2018-03-22 22:20:00
#> 624   0.697986263 2018-05-02 09:00:00
#> 625  -0.864349803 2018-06-11 19:40:00
#> 626  -1.091470351 2018-07-22 06:20:00
#> 627  -0.037051465 2018-08-31 17:00:00
#> 628   0.810053792 2018-10-11 03:40:00
#> 629  -0.499355412 2018-11-20 13:20:00
#> 630   0.948031588 2018-12-31 00:00:00
#> 631  -0.174245957 2018-01-01 00:00:00
#> 632  -1.106235952 2018-02-10 10:40:00
#> 633  -0.945985005 2018-03-22 22:20:00
#> 634   0.289089591 2018-05-02 09:00:00
#> 635   0.876913145 2018-06-11 19:40:00
#> 636  -1.148903940 2018-07-22 06:20:00
#> 637  -1.137612756 2018-08-31 17:00:00
#> 638  -1.437246735 2018-10-11 03:40:00
#> 639  -0.494143476 2018-11-20 13:20:00
#> 640   0.840801808 2018-12-31 00:00:00
#> 641   0.791534124 2018-01-01 00:00:00
#> 642  -0.168848948 2018-02-10 10:40:00
#> 643   0.612722104 2018-03-22 22:20:00
#> 644  -0.771158924 2018-05-02 09:00:00
#> 645   0.888628993 2018-06-11 19:40:00
#> 646   0.013214477 2018-07-22 06:20:00
#> 647   0.225339515 2018-08-31 17:00:00
#> 648  -0.729915210 2018-10-11 03:40:00
#> 649  -1.222487070 2018-11-20 13:20:00
#> 650   0.406805171 2018-12-31 00:00:00
#> 651  -0.751012223 2018-01-01 00:00:00
#> 652  -0.162116540 2018-02-10 10:40:00
#> 653   0.352010126 2018-03-22 22:20:00
#> 654  -0.289058300 2018-05-02 09:00:00
#> 655   0.104662227 2018-06-11 19:40:00
#> 656   0.720186531 2018-07-22 06:20:00
#> 657  -0.611046082 2018-08-31 17:00:00
#> 658  -1.106914072 2018-10-11 03:40:00
#> 659   0.534803326 2018-11-20 13:20:00
#> 660   0.736067968 2018-12-31 00:00:00
#> 661  -1.222501574 2018-01-01 00:00:00
#> 662   1.021415310 2018-02-10 10:40:00
#> 663   0.465165158 2018-03-22 22:20:00
#> 664   0.790472705 2018-05-02 09:00:00
#> 665  -0.130264801 2018-06-11 19:40:00
#> 666  -0.930285334 2018-07-22 06:20:00
#> 667  -0.364851004 2018-08-31 17:00:00
#> 668   0.153872493 2018-10-11 03:40:00
#> 669   0.413154818 2018-11-20 13:20:00
#> 670   2.480823360 2018-12-31 00:00:00
#> 671  -2.179956742 2018-01-01 00:00:00
#> 672   0.420874578 2018-02-10 10:40:00
#> 673  -0.357528325 2018-03-22 22:20:00
#> 674  -0.646861514 2018-05-02 09:00:00
#> 675  -0.050141801 2018-06-11 19:40:00
#> 676   0.416942847 2018-07-22 06:20:00
#> 677  -0.632587542 2018-08-31 17:00:00
#> 678   1.150146673 2018-10-11 03:40:00
#> 679  -0.235475907 2018-11-20 13:20:00
#> 680  -1.643107386 2018-12-31 00:00:00
#> 681  -1.503382146 2018-01-01 00:00:00
#> 682  -2.050584847 2018-02-10 10:40:00
#> 683  -0.753198229 2018-03-22 22:20:00
#> 684  -0.134141958 2018-05-02 09:00:00
#> 685   1.005782847 2018-06-11 19:40:00
#> 686   2.167186798 2018-07-22 06:20:00
#> 687   2.322556540 2018-08-31 17:00:00
#> 688  -1.020423391 2018-10-11 03:40:00
#> 689   0.048814436 2018-11-20 13:20:00
#> 690  -0.771888628 2018-12-31 00:00:00
#> 691  -0.785235068 2018-01-01 00:00:00
#> 692  -0.726603031 2018-02-10 10:40:00
#> 693   0.681878032 2018-03-22 22:20:00
#> 694  -0.229843287 2018-05-02 09:00:00
#> 695  -1.510601724 2018-06-11 19:40:00
#> 696  -0.583727687 2018-07-22 06:20:00
#> 697  -2.022918454 2018-08-31 17:00:00
#> 698   0.403504676 2018-10-11 03:40:00
#> 699   0.550015549 2018-11-20 13:20:00
#> 700   0.028357122 2018-12-31 00:00:00
#> 701   0.893165020 2018-01-01 00:00:00
#> 702  -0.376555496 2018-02-10 10:40:00
#> 703   0.605884808 2018-03-22 22:20:00
#> 704  -0.004874726 2018-05-02 09:00:00
#> 705  -0.520796373 2018-06-11 19:40:00
#> 706  -0.639018598 2018-07-22 06:20:00
#> 707  -0.635894137 2018-08-31 17:00:00
#> 708   0.106586975 2018-10-11 03:40:00
#> 709   1.176914248 2018-11-20 13:20:00
#> 710   0.447391153 2018-12-31 00:00:00
#> 711   2.272954766 2018-01-01 00:00:00
#> 712   0.136058206 2018-02-10 10:40:00
#> 713  -1.999039133 2018-03-22 22:20:00
#> 714  -0.420500870 2018-05-02 09:00:00
#> 715  -0.378407395 2018-06-11 19:40:00
#> 716   1.220774789 2018-07-22 06:20:00
#> 717  -1.541030292 2018-08-31 17:00:00
#> 718  -0.310310122 2018-10-11 03:40:00
#> 719  -0.020108184 2018-11-20 13:20:00
#> 720  -2.390200336 2018-12-31 00:00:00
#> 721   0.889865359 2018-01-01 00:00:00
#> 722  -1.482813325 2018-02-10 10:40:00
#> 723   0.445750348 2018-03-22 22:20:00
#> 724   1.369775856 2018-05-02 09:00:00
#> 725  -0.020110027 2018-06-11 19:40:00
#> 726  -0.109217587 2018-07-22 06:20:00
#> 727   0.264661745 2018-08-31 17:00:00
#> 728   0.303848264 2018-10-11 03:40:00
#> 729  -0.183388483 2018-11-20 13:20:00
#> 730   0.559649672 2018-12-31 00:00:00
#> 731  -0.186553842 2018-01-01 00:00:00
#> 732  -0.812275372 2018-02-10 10:40:00
#> 733  -1.640581672 2018-03-22 22:20:00
#> 734   0.507922478 2018-05-02 09:00:00
#> 735   1.754336961 2018-06-11 19:40:00
#> 736   0.592400202 2018-07-22 06:20:00
#> 737   1.016713288 2018-08-31 17:00:00
#> 738   0.121620586 2018-10-11 03:40:00
#> 739  -1.078067265 2018-11-20 13:20:00
#> 740  -1.143565720 2018-12-31 00:00:00
#> 741  -0.529643677 2018-01-01 00:00:00
#> 742  -0.681273156 2018-02-10 10:40:00
#> 743  -0.202447559 2018-03-22 22:20:00
#> 744   1.684495721 2018-05-02 09:00:00
#> 745  -1.033773237 2018-06-11 19:40:00
#> 746  -0.155976673 2018-07-22 06:20:00
#> 747  -0.046400637 2018-08-31 17:00:00
#> 748  -0.953628727 2018-10-11 03:40:00
#> 749   0.416260798 2018-11-20 13:20:00
#> 750   0.114029609 2018-12-31 00:00:00
#> 751   0.063918753 2018-01-01 00:00:00
#> 752  -0.919332238 2018-02-10 10:40:00
#> 753   0.901335290 2018-03-22 22:20:00
#> 754  -0.797728297 2018-05-02 09:00:00
#> 755   0.668221204 2018-06-11 19:40:00
#> 756   0.155214296 2018-07-22 06:20:00
#> 757   0.128688092 2018-08-31 17:00:00
#> 758  -1.533065453 2018-10-11 03:40:00
#> 759   0.202360673 2018-11-20 13:20:00
#> 760  -0.717538653 2018-12-31 00:00:00
#> 761   0.361694761 2018-01-01 00:00:00
#> 762   1.399004294 2018-02-10 10:40:00
#> 763   0.372698962 2018-03-22 22:20:00
#> 764  -1.565644294 2018-05-02 09:00:00
#> 765  -0.051694539 2018-06-11 19:40:00
#> 766   0.514082103 2018-07-22 06:20:00
#> 767   0.549899523 2018-08-31 17:00:00
#> 768   0.867816913 2018-10-11 03:40:00
#> 769   0.684360081 2018-11-20 13:20:00
#> 770  -0.162679977 2018-12-31 00:00:00
#> 771  -1.784364724 2018-01-01 00:00:00
#> 772  -1.037145566 2018-02-10 10:40:00
#> 773   0.830147723 2018-03-22 22:20:00
#> 774   0.607346945 2018-05-02 09:00:00
#> 775  -0.122186361 2018-06-11 19:40:00
#> 776   0.933125139 2018-07-22 06:20:00
#> 777  -0.961276680 2018-08-31 17:00:00
#> 778   0.255081707 2018-10-11 03:40:00
#> 779  -0.545401573 2018-11-20 13:20:00
#> 780   0.930360736 2018-12-31 00:00:00
#> 781  -0.537650558 2018-01-01 00:00:00
#> 782  -0.452426068 2018-02-10 10:40:00
#> 783  -0.439290972 2018-03-22 22:20:00
#> 784  -0.616223111 2018-05-02 09:00:00
#> 785   0.441634552 2018-06-11 19:40:00
#> 786   0.482597453 2018-07-22 06:20:00
#> 787   0.542144383 2018-08-31 17:00:00
#> 788  -2.290784653 2018-10-11 03:40:00
#> 789   0.310351853 2018-11-20 13:20:00
#> 790   1.404078979 2018-12-31 00:00:00
#> 791   1.377117639 2018-01-01 00:00:00
#> 792   1.060395652 2018-02-10 10:40:00
#> 793   0.632172891 2018-03-22 22:20:00
#> 794   1.084928035 2018-05-02 09:00:00
#> 795   1.356459439 2018-06-11 19:40:00
#> 796   0.362424040 2018-07-22 06:20:00
#> 797   2.169344465 2018-08-31 17:00:00
#> 798   0.139130512 2018-10-11 03:40:00
#> 799   1.376326526 2018-11-20 13:20:00
#> 800  -0.491449988 2018-12-31 00:00:00
#> 801   1.534915763 2018-01-01 00:00:00
#> 802  -0.416198718 2018-02-10 10:40:00
#> 803  -0.520543795 2018-03-22 22:20:00
#> 804   0.850583868 2018-05-02 09:00:00
#> 805   0.334496575 2018-06-11 19:40:00
#> 806  -0.829351774 2018-07-22 06:20:00
#> 807  -0.218695943 2018-08-31 17:00:00
#> 808  -1.545083718 2018-10-11 03:40:00
#> 809   0.233229783 2018-11-20 13:20:00
#> 810   0.031069642 2018-12-31 00:00:00
#> 811   0.357865645 2018-01-01 00:00:00
#> 812   1.608624223 2018-02-10 10:40:00
#> 813   1.429854263 2018-03-22 22:20:00
#> 814  -0.948339642 2018-05-02 09:00:00
#> 815   1.015705545 2018-06-11 19:40:00
#> 816   0.037734606 2018-07-22 06:20:00
#> 817  -1.744239400 2018-08-31 17:00:00
#> 818  -0.970073812 2018-10-11 03:40:00
#> 819  -0.229027360 2018-11-20 13:20:00
#> 820  -0.974845679 2018-12-31 00:00:00
#> 821   2.692372420 2018-01-01 00:00:00
#> 822   1.392393858 2018-02-10 10:40:00
#> 823   1.360076151 2018-03-22 22:20:00
#> 824  -0.365415875 2018-05-02 09:00:00
#> 825  -0.868937197 2018-06-11 19:40:00
#> 826  -0.506556729 2018-07-22 06:20:00
#> 827   1.214707304 2018-08-31 17:00:00
#> 828   0.506958680 2018-10-11 03:40:00
#> 829  -2.094971309 2018-11-20 13:20:00
#> 830   0.034079240 2018-12-31 00:00:00
#> 831   0.852728701 2018-01-01 00:00:00
#> 832   0.743228137 2018-02-10 10:40:00
#> 833   0.557153608 2018-03-22 22:20:00
#> 834  -1.248583225 2018-05-02 09:00:00
#> 835  -0.207874310 2018-06-11 19:40:00
#> 836   0.423969459 2018-07-22 06:20:00
#> 837  -0.506697440 2018-08-31 17:00:00
#> 838  -0.611523314 2018-10-11 03:40:00
#> 839   2.411659453 2018-11-20 13:20:00
#> 840  -0.164958139 2018-12-31 00:00:00
#> 841  -0.440406048 2018-01-01 00:00:00
#> 842   0.521976172 2018-02-10 10:40:00
#> 843  -1.918322250 2018-03-22 22:20:00
#> 844  -1.982649963 2018-05-02 09:00:00
#> 845   0.521200156 2018-06-11 19:40:00
#> 846   0.777783201 2018-07-22 06:20:00
#> 847  -0.812118870 2018-08-31 17:00:00
#> 848   0.910748320 2018-10-11 03:40:00
#> 849   0.948753954 2018-11-20 13:20:00
#> 850  -1.348049450 2018-12-31 00:00:00
#> 851   0.354175858 2018-01-01 00:00:00
#> 852   0.530263933 2018-02-10 10:40:00
#> 853  -0.310974933 2018-03-22 22:20:00
#> 854  -0.244155532 2018-05-02 09:00:00
#> 855  -0.291878192 2018-06-11 19:40:00
#> 856  -1.128086164 2018-07-22 06:20:00
#> 857  -1.122881586 2018-08-31 17:00:00
#> 858   2.053938640 2018-10-11 03:40:00
#> 859  -0.909518426 2018-11-20 13:20:00
#> 860   0.458379156 2018-12-31 00:00:00
#> 861  -0.046618452 2018-01-01 00:00:00
#> 862  -0.680954604 2018-02-10 10:40:00
#> 863  -1.488829402 2018-03-22 22:20:00
#> 864  -0.392513008 2018-05-02 09:00:00
#> 865  -1.749682017 2018-06-11 19:40:00
#> 866  -0.038667632 2018-07-22 06:20:00
#> 867  -0.797153244 2018-08-31 17:00:00
#> 868  -0.915920537 2018-10-11 03:40:00
#> 869   0.073267464 2018-11-20 13:20:00
#> 870   1.238171318 2018-12-31 00:00:00
#> 871   0.718371336 2018-01-01 00:00:00
#> 872   0.539466495 2018-02-10 10:40:00
#> 873  -1.061233995 2018-03-22 22:20:00
#> 874  -0.540737866 2018-05-02 09:00:00
#> 875  -0.715204714 2018-06-11 19:40:00
#> 876  -0.178920085 2018-07-22 06:20:00
#> 877   0.164970277 2018-08-31 17:00:00
#> 878  -0.727408166 2018-10-11 03:40:00
#> 879  -0.537837129 2018-11-20 13:20:00
#> 880  -0.599664904 2018-12-31 00:00:00
#> 881   1.182087754 2018-01-01 00:00:00
#> 882   0.189212875 2018-02-10 10:40:00
#> 883   0.248949114 2018-03-22 22:20:00
#> 884   1.046643411 2018-05-02 09:00:00
#> 885  -1.289300936 2018-06-11 19:40:00
#> 886   0.375115570 2018-07-22 06:20:00
#> 887  -0.556918391 2018-08-31 17:00:00
#> 888   0.302426630 2018-10-11 03:40:00
#> 889   0.222266466 2018-11-20 13:20:00
#> 890  -0.962166963 2018-12-31 00:00:00
#> 891   0.052032369 2018-01-01 00:00:00
#> 892   0.900040382 2018-02-10 10:40:00
#> 893  -0.394900880 2018-03-22 22:20:00
#> 894   0.954145996 2018-05-02 09:00:00
#> 895  -0.588214348 2018-06-11 19:40:00
#> 896  -1.476584527 2018-07-22 06:20:00
#> 897  -1.377775770 2018-08-31 17:00:00
#> 898  -1.345672313 2018-10-11 03:40:00
#> 899  -0.736637962 2018-11-20 13:20:00
#> 900  -0.470111496 2018-12-31 00:00:00
#> 901   1.380609338 2018-01-01 00:00:00
#> 902   1.675010928 2018-02-10 10:40:00
#> 903   1.176906634 2018-03-22 22:20:00
#> 904  -0.148898336 2018-05-02 09:00:00
#> 905  -0.177823364 2018-06-11 19:40:00
#> 906   0.803121858 2018-07-22 06:20:00
#> 907  -0.415795349 2018-08-31 17:00:00
#> 908   1.212489586 2018-10-11 03:40:00
#> 909   1.240360032 2018-11-20 13:20:00
#> 910   0.685675880 2018-12-31 00:00:00
#> 911  -0.026798678 2018-01-01 00:00:00
#> 912   0.309580509 2018-02-10 10:40:00
#> 913   0.249864328 2018-03-22 22:20:00
#> 914  -1.356460869 2018-05-02 09:00:00
#> 915   0.599379768 2018-06-11 19:40:00
#> 916   0.008647790 2018-07-22 06:20:00
#> 917   0.090716610 2018-08-31 17:00:00
#> 918  -0.657057410 2018-10-11 03:40:00
#> 919  -0.481782250 2018-11-20 13:20:00
#> 920   0.017757879 2018-12-31 00:00:00
#> 921  -0.859535756 2018-01-01 00:00:00
#> 922   1.357704977 2018-02-10 10:40:00
#> 923   1.215061177 2018-03-22 22:20:00
#> 924   1.453912917 2018-05-02 09:00:00
#> 925  -0.085911982 2018-06-11 19:40:00
#> 926  -0.617568750 2018-07-22 06:20:00
#> 927  -0.218260249 2018-08-31 17:00:00
#> 928  -1.326005211 2018-10-11 03:40:00
#> 929  -2.362208897 2018-11-20 13:20:00
#> 930  -1.409969554 2018-12-31 00:00:00
#> 931   0.254397136 2018-01-01 00:00:00
#> 932   0.295870299 2018-02-10 10:40:00
#> 933   0.056789794 2018-03-22 22:20:00
#> 934  -0.102654857 2018-05-02 09:00:00
#> 935   1.951923032 2018-06-11 19:40:00
#> 936   0.878565956 2018-07-22 06:20:00
#> 937  -1.306616928 2018-08-31 17:00:00
#> 938   0.090344272 2018-10-11 03:40:00
#> 939   0.902358040 2018-11-20 13:20:00
#> 940   0.549421324 2018-12-31 00:00:00
#> 941  -0.869839200 2018-01-01 00:00:00
#> 942  -0.039214648 2018-02-10 10:40:00
#> 943  -0.518451682 2018-03-22 22:20:00
#> 944  -0.911849617 2018-05-02 09:00:00
#> 945   0.150313640 2018-06-11 19:40:00
#> 946   0.440355449 2018-07-22 06:20:00
#> 947   1.319782059 2018-08-31 17:00:00
#> 948   0.246561126 2018-10-11 03:40:00
#> 949   0.941826599 2018-11-20 13:20:00
#> 950  -0.342151332 2018-12-31 00:00:00
#> 951  -0.276146811 2018-01-01 00:00:00
#> 952   0.383656780 2018-02-10 10:40:00
#> 953   1.446910225 2018-03-22 22:20:00
#> 954   1.733900291 2018-05-02 09:00:00
#> 955   0.456429376 2018-06-11 19:40:00
#> 956   0.707503490 2018-07-22 06:20:00
#> 957   2.063789224 2018-08-31 17:00:00
#> 958   0.023910586 2018-10-11 03:40:00
#> 959   0.245949461 2018-11-20 13:20:00
#> 960   0.272059145 2018-12-31 00:00:00
#> 961  -0.992458577 2018-01-01 00:00:00
#> 962  -0.027577951 2018-02-10 10:40:00
#> 963   2.222845163 2018-03-22 22:20:00
#> 964   0.155503899 2018-05-02 09:00:00
#> 965  -0.869116318 2018-06-11 19:40:00
#> 966  -1.174489450 2018-07-22 06:20:00
#> 967  -1.759677307 2018-08-31 17:00:00
#> 968   0.058361585 2018-10-11 03:40:00
#> 969   1.164519864 2018-11-20 13:20:00
#> 970   0.337627892 2018-12-31 00:00:00
#> 971  -1.054518720 2018-01-01 00:00:00
#> 972   0.660769107 2018-02-10 10:40:00
#> 973  -0.823409004 2018-03-22 22:20:00
#> 974   0.437033660 2018-05-02 09:00:00
#> 975   0.372388112 2018-06-11 19:40:00
#> 976  -1.646749128 2018-07-22 06:20:00
#> 977  -1.923552271 2018-08-31 17:00:00
#> 978   0.380833026 2018-10-11 03:40:00
#> 979   1.375705349 2018-11-20 13:20:00
#> 980   0.825848727 2018-12-31 00:00:00
#> 981  -0.416115003 2018-01-01 00:00:00
#> 982   0.982149586 2018-02-10 10:40:00
#> 983   0.242180727 2018-03-22 22:20:00
#> 984  -0.937926854 2018-05-02 09:00:00
#> 985   1.505858500 2018-06-11 19:40:00
#> 986  -1.065851174 2018-07-22 06:20:00
#> 987   0.162270527 2018-08-31 17:00:00
#> 988   0.260470392 2018-10-11 03:40:00
#> 989  -1.488201993 2018-11-20 13:20:00
#> 990   1.415183869 2018-12-31 00:00:00
#> 991   0.562507513 2018-01-01 00:00:00
#> 992   0.642055741 2018-02-10 10:40:00
#> 993  -0.890532644 2018-03-22 22:20:00
#> 994  -0.116537517 2018-05-02 09:00:00
#> 995  -0.941974751 2018-06-11 19:40:00
#> 996   1.115827918 2018-07-22 06:20:00
#> 997   0.470135134 2018-08-31 17:00:00
#> 998   0.860612712 2018-10-11 03:40:00
#> 999  -0.070396646 2018-11-20 13:20:00
#> 1000 -0.613180214 2018-12-31 00:00:00

## check output
identical(unique(dd), funique(dd))

#> [1] TRUE